+++ source/extensions/omni.isaac.robot_engine_bridge/plugins/Sensor/OccupancyGridMapComponent.cpp
@@ -1,4 +1,4 @@
-// Copyright (c) 2021, NVIDIA CORPORATION. All rights reserved.
+// Copyright (c) 2021-2022, NVIDIA CORPORATION. All rights reserved.
 //
 // NVIDIA CORPORATION and its licensors retain all intellectual property
 // and proprietary rights in and to this software, related documentation
@@ -113,7 +113,7 @@ void OccupancyGridMapComponent::publishAllMessages()
     if (mGenerator)
     {
         mGenerator->setTransform(trans, inputMinPoint, inputMaxPoint);
-        mGenerator->generate();
+        mGenerator->generate2d();
     }
     else
     {